requirements.txt
pyaudio
openai-whisper
groq
requests
playsound==1.2.2
